refactor(core/date-picker|time-picker): replace with rework implementation
