+++ packages/angular/src/declare-components.ts
@@ -20,7 +20,6 @@ export const DIRECTIVES = [
   d.IxCol,
   d.IxContentHeader,
   d.IxDateDropdown,
-  d.IxDatePicker,
   d.IxDatetimePicker,
   d.IxDivider,
   d.IxDrawer,
@@ -79,7 +78,6 @@ export const DIRECTIVES = [
   d.IxTabItem,
   d.IxTabs,
   d.IxTile,
-  d.IxTimePicker,
   d.IxToast,
   d.IxToastContainer,
   d.IxToggle,
