+++ packages/core/src/components.d.ts
@@ -19,8 +19,7 @@ import { ContentHeaderVariant } from "./components/content-header/content-header
 import { CssGridTemplateType } from "./components/css-grid/css-grid";
 import { DateDropdownOption, DateRangeChangeEvent } from "./components/date-dropdown/date-dropdown";
 import { DateTimeCardCorners } from "./components/date-time-card/date-time-card";
-import { DateChangeEvent, LegacyDateChangeEvent } from "./components/date-picker/date-picker";
-import { DateChangeEvent as DateChangeEvent1 } from "./components/date-picker-rework/date-picker-rework";
+import { DateChangeEvent } from "./components/date-picker/date-picker";
 import { DateTimeCardCorners as DateTimeCardCorners1 } from "./components/date-time-card/date-time-card";
 import { DateTimeDateChangeEvent, DateTimeSelectEvent } from "./components/datetime-picker/datetime-picker";
 import { DateTimeDateChangeEvent as DateTimeDateChangeEvent1, DateTimeSelectEvent as DateTimeSelectEvent1 } from "./components/datetime-picker-rework/datetime-picker-rework";
@@ -38,7 +37,6 @@ import { SliderMarker } from "./components/slider/slider";
 import { SplitButtonVariant } from "./components/split-button/split-button";
 import { TabClickDetail } from "./components/tab-item/tab-item";
 import { TimePickerCorners } from "./components/time-picker/time-picker";
-import { TimePickerCorners as TimePickerCorners1 } from "./components/time-picker-rework/time-picker-rework";
 import { ToastConfig, ToastType } from "./components/toast/toast-utils";
 import { TypedEvent } from "./components/utils/typed-event";
 import { TreeContext, TreeItemContext, TreeModel, UpdateCallback } from "./components/tree/tree-model";
@@ -58,8 +56,7 @@ export { ContentHeaderVariant } from "./components/content-header/content-header
 export { CssGridTemplateType } from "./components/css-grid/css-grid";
 export { DateDropdownOption, DateRangeChangeEvent } from "./components/date-dropdown/date-dropdown";
 export { DateTimeCardCorners } from "./components/date-time-card/date-time-card";
-export { DateChangeEvent, LegacyDateChangeEvent } from "./components/date-picker/date-picker";
-export { DateChangeEvent as DateChangeEvent1 } from "./components/date-picker-rework/date-picker-rework";
+export { DateChangeEvent } from "./components/date-picker/date-picker";
 export { DateTimeCardCorners as DateTimeCardCorners1 } from "./components/date-time-card/date-time-card";
 export { DateTimeDateChangeEvent, DateTimeSelectEvent } from "./components/datetime-picker/datetime-picker";
 export { DateTimeDateChangeEvent as DateTimeDateChangeEvent1, DateTimeSelectEvent as DateTimeSelectEvent1 } from "./components/datetime-picker-rework/datetime-picker-rework";
@@ -77,7 +74,6 @@ export { SliderMarker } from "./components/slider/slider";
 export { SplitButtonVariant } from "./components/split-button/split-button";
 export { TabClickDetail } from "./components/tab-item/tab-item";
 export { TimePickerCorners } from "./components/time-picker/time-picker";
-export { TimePickerCorners as TimePickerCorners1 } from "./components/time-picker-rework/time-picker-rework";
 export { ToastConfig, ToastType } from "./components/toast/toast-utils";
 export { TypedEvent } from "./components/utils/typed-event";
 export { TreeContext, TreeItemContext, TreeModel, UpdateCallback } from "./components/tree/tree-model";
@@ -551,55 +547,9 @@ export namespace Components {
         /**
           * Default behavior of the done event is to join the two events (date and time) into one combined string output. This combination can be configured over the delimiter
           * @since 1.1.0
+          * @deprecated Not used anymore see `this.dateChange`
          */
         "eventDelimiter": string;
-        /**
-          * Date format string. See {@link "https://moment.github.io/luxon/#/formatting?id=table-of-tokens"} for all available tokens.
-         */
-        "format": string;
-        /**
-          * Picker date. If the picker is in range mode this property is the start date. If set to `null` no default start date will be pre-selected.  Format is based on `format`
-          * @since 1.1.0
-         */
-        "from": string | null;
-        /**
-          * Get the current DateTime
-         */
-        "getCurrentDate": () => Promise<{ start: string; end: string; }>;
-        /**
-          * @deprecated Will be removed in 2.0.0
-         */
-        "individual": boolean;
-        /**
-          * The latest date that can be selected by the date picker. If not set there will be no restriction.
-          * @since 1.1.0
-         */
-        "maxDate": string;
-        /**
-          * The earliest date that can be selected by the date picker. If not set there will be no restriction.
-          * @since 1.1.0
-         */
-        "minDate": string;
-        /**
-          * If true a range of dates can be selected.
-         */
-        "range": boolean;
-        /**
-          * Text of date select button
-          * @since 1.1.0
-         */
-        "textSelectDate": string;
-        /**
-          * Picker date. If the picker is in range mode this property is the end date. If the picker is not in range mode leave this value `null`  Format is based on `format`
-          * @since 1.1.0
-         */
-        "to": string | null;
-    }
-    interface IxDatePickerRework {
-        /**
-          * Corner style
-         */
-        "corners": DateTimeCardCorners;
         /**
           * Date format string. See {@link "https://moment.github.io/luxon/#/formatting?id=table-of-tokens"} for all available tokens.
          */
@@ -613,6 +563,10 @@ export namespace Components {
           * Get the currently selected date-range.
          */
         "getCurrentDate": () => Promise<{ from: string; to: string; }>;
+        /**
+          * @deprecated Not supported since 2.0.0.
+         */
+        "individual": boolean;
         /**
           * Format of time string See {@link "https://moment.github.io/luxon/#/formatting?id=table-of-tokens"} for all available tokens.
          */
@@ -2019,16 +1973,16 @@ export namespace Components {
          */
         "corners": TimePickerCorners;
         /**
-          * Format of time string
+          * Format of time string See {@link "https://moment.github.io/luxon/#/formatting?id=table-of-tokens"} for all available tokens.
           * @since 1.1.0
          */
         "format": string;
         /**
-          * Get current time
+          * Get the current time based on the wanted format
          */
         "getCurrentTime": () => Promise<string>;
         /**
-          * @deprecated Will be removed in 2.0.0
+          * @deprecated Not supported since 2.0.0.
          */
         "individual": boolean;
         /**
@@ -2045,50 +1999,10 @@ export namespace Components {
         "showSeconds": boolean;
         /**
           * Show time reference input
-          * @since 1.1.0 time reference is default aligned with formt tt
+          * @since 1.1.0 time reference is default aligned with format tt
+          * @deprecated Since 2.0.0 time reference will be displayed depending on format.
          */
-        "showTimeReference": any;
-        /**
-          * Text of date select button
-          * @since 1.1.0
-         */
-        "textSelectTime": string;
-        /**
-          * Select time with format string
-          * @since 1.1.0
-         */
-        "time": string;
-        /**
-          * Set time reference
-         */
-        "timeReference": 'AM' | 'PM';
-    }
-    interface IxTimePickerRework {
-        /**
-          * Corner style
-         */
-        "corners": TimePickerCorners1;
-        /**
-          * Format of time string See {@link "https://moment.github.io/luxon/#/formatting?id=table-of-tokens"} for all available tokens.
-          * @since 1.1.0
-         */
-        "format": string;
-        /**
-          * Get the current time based on the wanted format
-         */
-        "getCurrentTime": () => Promise<string>;
-        /**
-          * Show hour input
-         */
-        "showHours": boolean;
-        /**
-          * Show minutes input
-         */
-        "showMinutes": boolean;
-        /**
-          * Show seconds input
-         */
-        "showSeconds": boolean;
+        "showTimeReference"?: boolean;
         /**
           * Controls the visual presentation and styling of the component when it is displayed as a standalone element
          */
@@ -2098,6 +2012,10 @@ export namespace Components {
           * @since 1.1.0
          */
         "textSelectTime": string;
+        /**
+          * Text for top label
+         */
+        "textTime": string;
         /**
           * Select time with format string Format has to match the `format` property.
           * @since 1.1.0
@@ -2445,10 +2363,6 @@ export interface IxDatePickerCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLIxDatePickerElement;
 }
-export interface IxDatePickerReworkCustomEvent<T> extends CustomEvent<T> {
-    detail: T;
-    target: HTMLIxDatePickerReworkElement;
-}
 export interface IxDatetimePickerCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLIxDatetimePickerElement;
@@ -2580,10 +2494,6 @@ export interface IxTabsCustomEvent<T> extends CustomEvent<T> {
 export interface IxTimePickerCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLIxTimePickerElement;
-}
-export interface IxTimePickerReworkCustomEvent<T> extends CustomEvent<T> {
-    detail: T;
-    target: HTMLIxTimePickerReworkElement;
 }
 export interface IxToastCustomEvent<T> extends CustomEvent<T> {
     detail: T;
@@ -2794,12 +2704,6 @@ declare global {
     var HTMLIxDatePickerElement: {
         prototype: HTMLIxDatePickerElement;
         new (): HTMLIxDatePickerElement;
-    };
-    interface HTMLIxDatePickerReworkElement extends Components.IxDatePickerRework, HTMLStencilElement {
-    }
-    var HTMLIxDatePickerReworkElement: {
-        prototype: HTMLIxDatePickerReworkElement;
-        new (): HTMLIxDatePickerReworkElement;
     };
     interface HTMLIxDateTimeCardElement extends Components.IxDateTimeCard, HTMLStencilElement {
     }
@@ -3230,12 +3134,6 @@ declare global {
         prototype: HTMLIxTimePickerElement;
         new (): HTMLIxTimePickerElement;
     };
-    interface HTMLIxTimePickerReworkElement extends Components.IxTimePickerRework, HTMLStencilElement {
-    }
-    var HTMLIxTimePickerReworkElement: {
-        prototype: HTMLIxTimePickerReworkElement;
-        new (): HTMLIxTimePickerReworkElement;
-    };
     interface HTMLIxToastElement extends Components.IxToast, HTMLStencilElement {
     }
     var HTMLIxToastElement: {
@@ -3349,7 +3247,6 @@ declare global {
         "ix-css-grid-item": HTMLIxCssGridItemElement;
         "ix-date-dropdown": HTMLIxDateDropdownElement;
         "ix-date-picker": HTMLIxDatePickerElement;
-        "ix-date-picker-rework": HTMLIxDatePickerReworkElement;
         "ix-date-time-card": HTMLIxDateTimeCardElement;
         "ix-datetime-picker": HTMLIxDatetimePickerElement;
         "ix-datetime-picker-rework": HTMLIxDatetimePickerReworkElement;
@@ -3412,7 +3309,6 @@ declare global {
         "ix-tabs": HTMLIxTabsElement;
         "ix-tile": HTMLIxTileElement;
         "ix-time-picker": HTMLIxTimePickerElement;
-        "ix-time-picker-rework": HTMLIxTimePickerReworkElement;
         "ix-toast": HTMLIxToastElement;
         "ix-toast-container": HTMLIxToastContainerElement;
         "ix-toggle": HTMLIxToggleElement;
@@ -3947,71 +3843,9 @@ declare namespace LocalJSX {
         /**
           * Default behavior of the done event is to join the two events (date and time) into one combined string output. This combination can be configured over the delimiter
           * @since 1.1.0
+          * @deprecated Not used anymore see `this.dateChange`
          */
         "eventDelimiter"?: string;
-        /**
-          * Date format string. See {@link "https://moment.github.io/luxon/#/formatting?id=table-of-tokens"} for all available tokens.
-         */
-        "format"?: string;
-        /**
-          * Picker date. If the picker is in range mode this property is the start date. If set to `null` no default start date will be pre-selected.  Format is based on `format`
-          * @since 1.1.0
-         */
-        "from"?: string | null;
-        /**
-          * @deprecated Will be removed in 2.0.0
-         */
-        "individual"?: boolean;
-        /**
-          * The latest date that can be selected by the date picker. If not set there will be no restriction.
-          * @since 1.1.0
-         */
-        "maxDate"?: string;
-        /**
-          * The earliest date that can be selected by the date picker. If not set there will be no restriction.
-          * @since 1.1.0
-         */
-        "minDate"?: string;
-        /**
-          * Date change event  If datepicker is in range mode the event detail will be sperated with a `-` e.g. `2022/10/22 - 2022/10/24` (start and end). If range mode is chosen consider to use `dateRangeChange`.
-          * @deprecated String output will be removed. Set ´doneEventDelimiter´ to undefined or null to get date change object instead of a string
-         */
-        "onDateChange"?: (event: IxDatePickerCustomEvent<LegacyDateChangeEvent>) => void;
-        /**
-          * Date range change. Only triggered if datepicker is in range mode
-          * @since 1.1.0
-         */
-        "onDateRangeChange"?: (event: IxDatePickerCustomEvent<DateChangeEvent>) => void;
-        /**
-          * Date selection confirmed via button action
-          * @since 1.1.0
-         */
-        "onDateSelect"?: (event: IxDatePickerCustomEvent<DateChangeEvent>) => void;
-        /**
-          * Date selection confirmed via button action
-          * @deprecated Will be removed in 2.0.0. Use `dateSelect`
-         */
-        "onDone"?: (event: IxDatePickerCustomEvent<string>) => void;
-        /**
-          * If true a range of dates can be selected.
-         */
-        "range"?: boolean;
-        /**
-          * Text of date select button
-          * @since 1.1.0
-         */
-        "textSelectDate"?: string;
-        /**
-          * Picker date. If the picker is in range mode this property is the end date. If the picker is not in range mode leave this value `null`  Format is based on `format`
-          * @since 1.1.0
-         */
-        "to"?: string | null;
-    }
-    interface IxDatePickerRework {
-        /**
-          * Corner style
-         */
-        "corners"?: DateTimeCardCorners;
         /**
           * Date format string. See {@link "https://moment.github.io/luxon/#/formatting?id=table-of-tokens"} for all available tokens.
          */
@@ -4021,6 +3855,10 @@ declare namespace LocalJSX {
           * @since 1.1.0
          */
         "from"?: string | undefined;
+        /**
+          * @deprecated Not supported since 2.0.0.
+         */
+        "individual"?: boolean;
         /**
           * Format of time string See {@link "https://moment.github.io/luxon/#/formatting?id=table-of-tokens"} for all available tokens.
          */
@@ -4036,21 +3874,25 @@ declare namespace LocalJSX {
          */
         "minDate"?: string;
         /**
-          * Triggers if the date selection changes.
+          * Triggers if the date selection changes. Note: Since 2.0.0 `dateChange` does not dispatch detail property as `string`
           * @since 2.0.0
          */
-        "onDateChange"?: (event: IxDatePickerReworkCustomEvent<DateChangeEvent1>) => void;
+        "onDateChange"?: (event: IxDatePickerCustomEvent<DateChangeEvent>) => void;
         /**
           * Triggers if the date selection changes. Only triggered if date-picker-rework is in range mode.
-          * @since 1.1.0
-          * @deprecated Use `dateChange` (triggers on both modes)
+          * @since 2.0.0
          */
-        "onDateRangeChange"?: (event: IxDatePickerReworkCustomEvent<DateChangeEvent1>) => void;
+        "onDateRangeChange"?: (event: IxDatePickerCustomEvent<DateChangeEvent>) => void;
         /**
           * Date selection confirmed via button action
           * @since 1.1.0
          */
-        "onDateSelect"?: (event: IxDatePickerReworkCustomEvent<DateChangeEvent1>) => void;
+        "onDateSelect"?: (event: IxDatePickerCustomEvent<DateChangeEvent>) => void;
+        /**
+          * Date selection confirmed via button action
+          * @deprecated NOT getting dispatched after 2.0.0. Use `dateSelect`.
+         */
+        "onDone"?: (event: IxDatePickerCustomEvent<string>) => void;
         /**
           * If true a date-range can be selected (from/to).
          */
@@ -5581,22 +5423,27 @@ declare namespace LocalJSX {
          */
         "corners"?: TimePickerCorners;
         /**
-          * Format of time string
+          * Format of time string See {@link "https://moment.github.io/luxon/#/formatting?id=table-of-tokens"} for all available tokens.
           * @since 1.1.0
          */
         "format"?: string;
         /**
-          * @deprecated Will be removed in 2.0.0
+          * @deprecated Not supported since 2.0.0.
          */
         "individual"?: boolean;
         /**
           * Time event
+          * @deprecated Will be removed in 3.0.0. Use `time-select` event.
          */
         "onDone"?: (event: IxTimePickerCustomEvent<string>) => void;
         /**
           * Time change event
          */
         "onTimeChange"?: (event: IxTimePickerCustomEvent<string>) => void;
+        /**
+          * Time event
+         */
+        "onTimeSelect"?: (event: IxTimePickerCustomEvent<string>) => void;
         /**
           * Show hour input
          */
@@ -5611,54 +5458,10 @@ declare namespace LocalJSX {
         "showSeconds"?: boolean;
         /**
           * Show time reference input
-          * @since 1.1.0 time reference is default aligned with formt tt
+          * @since 1.1.0 time reference is default aligned with format tt
+          * @deprecated Since 2.0.0 time reference will be displayed depending on format.
          */
-        "showTimeReference"?: any;
-        /**
-          * Text of date select button
-          * @since 1.1.0
-         */
-        "textSelectTime"?: string;
-        /**
-          * Select time with format string
-          * @since 1.1.0
-         */
-        "time"?: string;
-        /**
-          * Set time reference
-         */
-        "timeReference"?: 'AM' | 'PM';
-    }
-    interface IxTimePickerRework {
-        /**
-          * Corner style
-         */
-        "corners"?: TimePickerCorners1;
-        /**
-          * Format of time string See {@link "https://moment.github.io/luxon/#/formatting?id=table-of-tokens"} for all available tokens.
-          * @since 1.1.0
-         */
-        "format"?: string;
-        /**
-          * Time change event
-         */
-        "onTimeChange"?: (event: IxTimePickerReworkCustomEvent<string>) => void;
-        /**
-          * Time event
-         */
-        "onTimeSelect"?: (event: IxTimePickerReworkCustomEvent<string>) => void;
-        /**
-          * Show hour input
-         */
-        "showHours"?: boolean;
-        /**
-          * Show minutes input
-         */
-        "showMinutes"?: boolean;
-        /**
-          * Show seconds input
-         */
-        "showSeconds"?: boolean;
+        "showTimeReference"?: boolean;
         /**
           * Controls the visual presentation and styling of the component when it is displayed as a standalone element
          */
@@ -5668,6 +5471,10 @@ declare namespace LocalJSX {
           * @since 1.1.0
          */
         "textSelectTime"?: string;
+        /**
+          * Text for top label
+         */
+        "textTime"?: string;
         /**
           * Select time with format string Format has to match the `format` property.
           * @since 1.1.0
@@ -6035,7 +5842,6 @@ declare namespace LocalJSX {
         "ix-css-grid-item": IxCssGridItem;
         "ix-date-dropdown": IxDateDropdown;
         "ix-date-picker": IxDatePicker;
-        "ix-date-picker-rework": IxDatePickerRework;
         "ix-date-time-card": IxDateTimeCard;
         "ix-datetime-picker": IxDatetimePicker;
         "ix-datetime-picker-rework": IxDatetimePickerRework;
@@ -6098,7 +5904,6 @@ declare namespace LocalJSX {
         "ix-tabs": IxTabs;
         "ix-tile": IxTile;
         "ix-time-picker": IxTimePicker;
-        "ix-time-picker-rework": IxTimePickerRework;
         "ix-toast": IxToast;
         "ix-toast-container": IxToastContainer;
         "ix-toggle": IxToggle;
@@ -6170,7 +5975,6 @@ declare module "@stencil/core" {
              */
             "ix-date-dropdown": LocalJSX.IxDateDropdown & JSXBase.HTMLAttributes<HTMLIxDateDropdownElement>;
             "ix-date-picker": LocalJSX.IxDatePicker & JSXBase.HTMLAttributes<HTMLIxDatePickerElement>;
-            "ix-date-picker-rework": LocalJSX.IxDatePickerRework & JSXBase.HTMLAttributes<HTMLIxDatePickerReworkElement>;
             "ix-date-time-card": LocalJSX.IxDateTimeCard & JSXBase.HTMLAttributes<HTMLIxDateTimeCardElement>;
             "ix-datetime-picker": LocalJSX.IxDatetimePicker & JSXBase.HTMLAttributes<HTMLIxDatetimePickerElement>;
             "ix-datetime-picker-rework": LocalJSX.IxDatetimePickerRework & JSXBase.HTMLAttributes<HTMLIxDatetimePickerReworkElement>;
@@ -6290,7 +6094,6 @@ declare module "@stencil/core" {
             "ix-tabs": LocalJSX.IxTabs & JSXBase.HTMLAttributes<HTMLIxTabsElement>;
             "ix-tile": LocalJSX.IxTile & JSXBase.HTMLAttributes<HTMLIxTileElement>;
             "ix-time-picker": LocalJSX.IxTimePicker & JSXBase.HTMLAttributes<HTMLIxTimePickerElement>;
-            "ix-time-picker-rework": LocalJSX.IxTimePickerRework & JSXBase.HTMLAttributes<HTMLIxTimePickerReworkElement>;
             "ix-toast": LocalJSX.IxToast & JSXBase.HTMLAttributes<HTMLIxToastElement>;
             "ix-toast-container": LocalJSX.IxToastContainer & JSXBase.HTMLAttributes<HTMLIxToastContainerElement>;
             "ix-toggle": LocalJSX.IxToggle & JSXBase.HTMLAttributes<HTMLIxToggleElement>;
