+++ packages/core/src/components/date-picker/date-picker.tsx
@@ -18,30 +18,35 @@ import {
   Method,
   Prop,
   State,
+  Watch,
 } from '@stencil/core';
-import { DateTime, Info, MonthNumbers } from 'luxon';
 import { DateTimeCardCorners } from '../date-time-card/date-time-card';
+
+import { DateTime, Info } from 'luxon';
+import { OnListener } from '../utils/listener';
 
 export type DateChangeEvent = {
   from: string;
   to: string;
 };
 
-export type LegacyDateChangeEvent = DateChangeEvent | string;
-
 export type DateTimeCorners = DateTimeCardCorners;
 
+interface CalendarWeek {
+  weekNumber: number;
+  dayNumbers: number[];
+}
+
+/**
+ * @internal
+ */
 @Component({
   tag: 'ix-date-picker',
-  styleUrl: './date-picker.scss',
+  styleUrl: 'date-picker.scss',
   shadow: true,
 })
 export class DatePicker {
   @Element() hostElement: HTMLIxDatePickerElement;
-
-  private daysInWeek = 7;
-  private dayNames = Info.weekdays();
-  private monthNames = Info.months();
 
   /**
    * Date format string.
@@ -50,14 +55,9 @@ export class DatePicker {
   @Prop() format: string = 'yyyy/LL/dd';
 
   /**
-   * If true a range of dates can be selected.
+   * If true a date-range can be selected (from/to).
    */
   @Prop() range: boolean = true;
-
-  /**
-   * @deprecated Will be removed in 2.0.0
-   */
-  @Prop() individual: boolean = true;
 
   /**
    * Corner style
@@ -65,24 +65,44 @@ export class DatePicker {
   @Prop() corners: DateTimeCardCorners = 'rounded';
 
   /**
-   * Picker date. If the picker is in range mode this property is the start date.
-   * If set to `null` no default start date will be pre-selected.
-   *
-   * Format is based on `format`
+   * The selected starting date. If the date-picker-rework is not in range mode this is the selected date.
+   * Format has to match the `format` property.
    *
    * @since 1.1.0
    */
-  @Prop() from: string | null = DateTime.now().toFormat(this.format);
+  @Prop() from: string | undefined;
+
+  @Watch('from')
+  watchFromPropHandler(newValue: string) {
+    this.currFromDate = newValue
+      ? DateTime.fromFormat(newValue, this.format)
+      : undefined;
+
+    if (this.currFromDate?.isValid) {
+      this.selectedYear = this.currFromDate.year;
+      this.selectedMonth = this.currFromDate.month - 1;
+    }
+  }
 
   /**
-   * Picker date. If the picker is in range mode this property is the end date.
-   * If the picker is not in range mode leave this value `null`
-   *
-   * Format is based on `format`
+   * The selected end date. If the the date-picker-rework is not in range mode this property has no impact.
+   * Format has to match the `format` property.
    *
    * @since 1.1.0
    */
-  @Prop() to: string | null = null;
+  @Prop() to: string | undefined;
+
+  @Watch('to')
+  watchToPropHandler(newValue: string) {
+    this.currToDate = newValue
+      ? DateTime.fromFormat(newValue, this.format)
+      : undefined;
+
+    if (this.currToDate?.isValid) {
+      this.selectedYear = this.currToDate.year;
+      this.selectedMonth = this.currToDate.month - 1;
+    }
+  }
 
   /**
    * The earliest date that can be selected by the date picker.
@@ -101,65 +121,53 @@ export class DatePicker {
   @Prop() maxDate: string;
 
   /**
-   * Default behavior of the done event is to join the two events (date and time) into one combined string output.
-   * This combination can be configured over the delimiter
-   *
-   * @since 1.1.0
-   */
-  @Prop() eventDelimiter = ' - ';
-
-  /**
-   * Text of date select button
+   * Text of the button that confirms date selection.
    *
    * @since 1.1.0
    */
   @Prop() textSelectDate = 'Done';
 
-  @State() yearValue = this.year;
-  @State() today = DateTime.now();
-  @State() monthValue: number = this.month;
-  @State() calendar: [number, number[]][] = [];
-
-  @State() years = [...Array(10).keys()].map((year) => year + this.year - 5);
-  @State() tempYear: number = this.yearValue;
-  @State() tempMonth: number = this.monthValue;
-  @State() start: DateTime = DateTime.fromObject({
-    year: this.year,
-    month: this.month,
-    day: this.day,
-  });
-
-  @State() end: DateTime = this.to
-    ? DateTime.fromFormat(this.to, this.format)
-    : null;
-
-  @State() dropdownButtonRef: HTMLElement;
-  @State() yearContainerRef: HTMLElement;
-
   /**
-   * Date change event
+   * The index of which day to start the week on, based on the Locale#weekdays array.
+   * E.g. if the locale is en-us, weekStartIndex = 1 results in starting the week on monday.
    *
-   * If datepicker is in range mode the event detail will be sperated with a `-` e.g.
-   * `2022/10/22 - 2022/10/24` (start and end). If range mode is chosen consider to use `dateRangeChange`.
-   *
-   * @deprecated String output will be removed. Set ´doneEventDelimiter´ to undefined or null to get date change object instead of a string
+   * @since 2.0.0
    */
-  @Event() dateChange: EventEmitter<LegacyDateChangeEvent>;
+  @Prop() weekStartIndex = 0;
 
   /**
-   * Date range change.
-   * Only triggered if datepicker is in range mode
+   * Format of time string
+   * See {@link "https://moment.github.io/luxon/#/formatting?id=table-of-tokens"} for all available tokens.
+   */
+  @Prop() locale: string = undefined;
+  @Watch('locale')
+  onLocaleChange() {
+    this.setTranslations();
+  }
+
+  /**
+   * @deprecated Not supported since 2.0.0.
+   */
+  @Prop() individual: boolean = true;
+
+  /** @internal */
+  @Prop() standaloneAppearance = true;
+
+  /**
+   * Triggers if the date selection changes.
+   * Note: Since 2.0.0 `dateChange` does not dispatch detail property as `string`
    *
-   * @since 1.1.0
+   * @since 2.0.0
+   */
+  @Event() dateChange: EventEmitter<DateChangeEvent>;
+
+  /**
+   * Triggers if the date selection changes.
+   * Only triggered if date-picker-rework is in range mode.
+   *
+   * @since 2.0.0
    */
   @Event() dateRangeChange: EventEmitter<DateChangeEvent>;
-
-  /**
-   * Date selection confirmed via button action
-   *
-   * @deprecated Will be removed in 2.0.0. Use `dateSelect`
-   */
-  @Event() done: EventEmitter<string>;
 
   /**
    * Date selection confirmed via button action
@@ -168,152 +176,286 @@ export class DatePicker {
    */
   @Event() dateSelect: EventEmitter<DateChangeEvent>;
 
-  get year() {
-    if (this.from !== null) {
-      return DateTime.fromFormat(this.from, this.format).year;
+  /**
+   * Date selection confirmed via button action
+   *
+   * @deprecated NOT getting dispatched after 2.0.0. Use `dateSelect`.
+   */
+  @Event() done: EventEmitter<string>;
+
+  /**
+   * Default behavior of the done event is to join the two events (date and time) into one combined string output.
+   * This combination can be configured over the delimiter
+   *
+   * @since 1.1.0
+   * @deprecated Not used anymore see `this.dateChange`
+   */
+  @Prop() eventDelimiter = ' - ';
+
+  /**
+   * Get the currently selected date-range.
+   */
+  @Method()
+  async getCurrentDate() {
+    const _from = this.currFromDate?.isValid
+      ? this.currFromDate?.toFormat(this.format)
+      : undefined;
+    const _to = this.currToDate?.isValid
+      ? this.currToDate?.toFormat(this.format)
+      : undefined;
+
+    if (this.range) {
+      return {
+        from: _from,
+        to: _to,
+      };
     }
 
-    return DateTime.now().year;
+    return {
+      from: _from,
+      to: undefined,
+    };
   }
 
-  get day() {
-    if (this.from !== null) {
-      return DateTime.fromFormat(this.from, this.format).day;
+  @State() currFromDate: DateTime;
+  @State() currToDate: DateTime;
+
+  @State() selectedYear: number;
+  @State() tempYear: number;
+  @State() startYear: number;
+  @State() endYear: number;
+  @State() selectedMonth: number;
+  @State() tempMonth: number;
+
+  @State() dropdownButtonRef: HTMLElement;
+  @State() yearContainerRef: HTMLElement;
+
+  @State() dayNames: string[];
+  @State() monthNames: string[];
+  @State() firstMonthRef: HTMLElement;
+  @State() focusedDay: number = 1;
+  @State() focusedDayElem: HTMLElement;
+
+  private isDayFocus: boolean;
+  private monthChangedFromFocus: boolean;
+  private readonly DAYS_IN_WEEK = 7;
+  private calendar: CalendarWeek[];
+
+  @OnListener<DatePicker>('keydown')
+  handleKeyUp(event: KeyboardEvent) {
+    if (!this.isDayFocus) {
+      return;
     }
 
-    return null;
-  }
+    let _focusedDay = this.focusedDay;
 
-  get month() {
-    if (this.from !== null) {
-      return DateTime.fromFormat(this.from, this.format).month;
+    switch (event.key) {
+      case 'ArrowLeft':
+        _focusedDay--;
+        break;
+      case 'ArrowRight':
+        _focusedDay++;
+        break;
+      case 'ArrowUp':
+        _focusedDay = _focusedDay - 7;
+        break;
+      case 'ArrowDown':
+        _focusedDay = _focusedDay + 7;
+        break;
+      default:
+        return;
     }
 
-    return DateTime.now().month;
+    if (_focusedDay > this.getDaysInCurrentMonth()) {
+      _focusedDay = _focusedDay - this.getDaysInCurrentMonth();
+      this.changeToAdjacentMonth(1);
+      this.monthChangedFromFocus = true;
+    } else if (_focusedDay < 1) {
+      this.changeToAdjacentMonth(-1);
+      _focusedDay = _focusedDay + this.getDaysInCurrentMonth();
+      this.monthChangedFromFocus = true;
+    }
+
+    this.focusedDay = _focusedDay;
   }
 
-  private onDone() {
-    this.done.emit(this.getOutputFormat());
+  private getDaysInCurrentMonth(): number {
+    return DateTime.utc(this.selectedYear, this.selectedMonth + 1).daysInMonth;
+  }
 
-    this.dateSelect.emit({
-      from: this.start?.toFormat(this.format),
-      to: this.end?.toFormat(this.format),
+  onDayBlur() {
+    this.isDayFocus = false;
+  }
+
+  onDayFocus() {
+    this.isDayFocus = true;
+  }
+
+  componentWillLoad() {
+    this.setTranslations();
+
+    this.currFromDate = this.from
+      ? DateTime.fromFormat(this.from, this.format)
+      : undefined;
+    this.currToDate = this.to
+      ? DateTime.fromFormat(this.to, this.format)
+      : undefined;
+
+    const year = this.currFromDate?.year ?? DateTime.now().year;
+    this.startYear = year - 5;
+    this.endYear = year + 5;
+
+    this.selectedMonth = (this.currFromDate?.month ?? DateTime.now().month) - 1;
+    this.selectedYear = year;
+    this.tempMonth = this.selectedMonth;
+    this.tempYear = this.selectedYear;
+  }
+
+  componentWillRender() {
+    this.calculateCalendar();
+  }
+
+  componentDidRender() {
+    if (!this.monthChangedFromFocus && !this.isDayFocus) {
+      return;
+    }
+
+    const dayElem = this.hostElement.shadowRoot.querySelector(
+      `[id=day-cell-${this.focusedDay}]`
+    ) as HTMLElement;
+    dayElem.focus();
+  }
+
+  private setTranslations() {
+    this.dayNames = this.rotateWeekDayNames(
+      Info.weekdays('long', {
+        locale: this.locale,
+      }),
+      this.weekStartIndex
+    );
+
+    this.monthNames = Info.months('long', {
+      locale: this.locale,
     });
   }
 
-  private onDateChange() {
-    const from = this.start?.toFormat(this.format);
-    const to = this.end?.toFormat(this.format);
+  /**
+   * Rotate the WeekdayNames array.
+   * Based on the position that should be the new 0-index.
+   */
+  private rotateWeekDayNames(weekdays: string[], index: number): string[] {
+    const clone = [...weekdays];
 
-    this.from = from;
-    this.to = to;
-
-    if (this.eventDelimiter) {
-      this.dateChange.emit(this.getOutputFormat());
-    } else {
-      this.dateChange.emit({
-        from,
-        to,
-      });
+    if (index === 0) {
+      return clone;
     }
 
-    if (this.range) {
-      this.dateRangeChange.emit({
-        from,
-        to,
-      });
-    }
+    index = -index;
+    const len = weekdays.length;
+
+    clone.push(...clone.splice(0, ((-index % len) + len) % len));
+    return clone;
   }
 
-  private getStartOfMonth(
-    year = DateTime.local().get('year'),
-    month = DateTime.local().get('month')
-  ) {
-    return DateTime.local(year, month).startOf('month');
-  }
-
-  private getEndOfMonth(
-    year = DateTime.local().get('year'),
-    month = DateTime.local().get('month')
-  ) {
-    return DateTime.local(year, month).endOf('month');
-  }
-
-  private getDaysInMonth(
-    start = this.getStartOfMonth(),
-    end = this.getEndOfMonth()
-  ) {
-    return Math.ceil(end.diff(start, 'days').days);
+  private async onDone() {
+    const date = await this.getCurrentDate();
+    this.dateSelect.emit(date);
   }
 
   private calculateCalendar() {
-    const start = this.getStartOfMonth(this.yearValue, this.monthValue);
-    const end = this.getEndOfMonth(this.yearValue, this.monthValue);
-    const totalDays = this.getDaysInMonth(start, end);
-    const totalWeeks = 6;
-    const totalDaysInWeeks = totalWeeks * this.daysInWeek;
-    const startWeekDay = start.weekday;
-    const prependDays = startWeekDay - 1;
-    const appendDays = totalDaysInWeeks - totalDays - prependDays;
-    let weekdays: number[][] = [];
-    const calendar: [number, number[]][] = [];
+    const calendar: CalendarWeek[] = [];
+    const month = DateTime.utc(this.selectedYear, this.selectedMonth + 1);
+    const monthStart = month.startOf('month');
+    const monthEnd = month.endOf('month');
+    let startWeek = monthStart.weekNumber;
+    let endWeek = monthEnd.weekNumber;
+    let monthStartWeekDayIndex = monthStart.weekday - 1;
+    let monthEndWeekDayIndex = monthEnd.weekday - 1;
 
-    // create list of days
-    let days = [...new Array(totalDaysInWeeks).keys()].map((day) => day + 1);
+    if (this.weekStartIndex !== 0) {
+      // Find the positions where to start/stop counting the day-numbers based on which day the week starts
+      const weekdays = Info.weekdays();
+      const monthStartWeekDayName = weekdays[monthStart.weekday];
 
-    // add start empty days
-    days.unshift(...new Array(prependDays));
+      monthStartWeekDayIndex = this.dayNames.findIndex(
+        (d) => d === monthStartWeekDayName
+      );
+      const monthEndWeekDayName = weekdays[monthEnd.weekday];
+      monthEndWeekDayIndex = this.dayNames.findIndex(
+        (d) => d === monthEndWeekDayName
+      );
+    }
 
-    // remove & add end days
-    days = days.slice(0, days.length - prependDays - appendDays);
-    days.push(...new Array(appendDays));
+    let correctLastWeek = false;
+    if (endWeek === 1) {
+      endWeek = monthEnd.weeksInWeekYear + 1;
+      correctLastWeek = true;
+    }
 
-    // make weeks
-    weekdays = days.reduce((result, item, index) => {
-      const weekIndex = Math.floor(index / this.daysInWeek);
+    let correctFirstWeek = false;
+    if (startWeek === monthStart.weeksInWeekYear) {
+      startWeek = 1;
+      endWeek++;
 
-      if (!result[weekIndex]) result[weekIndex] = [];
+      correctFirstWeek = true;
+    }
 
-      result[weekIndex].push(item);
+    let currDayNumber = 1;
+    for (
+      let weekIndex = startWeek;
+      weekIndex <= endWeek && currDayNumber <= 31;
+      weekIndex++
+    ) {
+      const daysArr: number[] = [];
 
-      return result;
-    }, []);
+      for (let j = 0; j < this.DAYS_IN_WEEK && currDayNumber <= 31; j++) {
+        // Display empty cells until the calender starts/has ended
+        if (
+          (weekIndex === startWeek && j < monthStartWeekDayIndex) ||
+          (weekIndex === endWeek && j > monthEndWeekDayIndex)
+        ) {
+          daysArr.push(undefined);
+        } else {
+          daysArr.push(currDayNumber++);
+        }
+      }
 
-    for (let index = 1; index <= totalWeeks; index++) {
-      const week = weekdays[index - 1];
-      const firstWeekDay = week.find((day) => day !== undefined);
-      const weekNumber = firstWeekDay
-        ? DateTime.local(
-            this.yearValue,
-            this.monthValue,
-            weekdays[index - 1][0]
-          ).weekNumber
-        : undefined;
-      calendar.push([weekNumber, week]);
+      if (correctFirstWeek || correctLastWeek) {
+        if (weekIndex === 1) {
+          calendar.push({
+            weekNumber: monthStart.weeksInWeekYear,
+            dayNumbers: daysArr,
+          });
+        } else if (weekIndex === monthEnd.weekNumber) {
+          calendar.push({
+            weekNumber: 1,
+            dayNumbers: daysArr,
+          });
+        } else {
+          calendar.push({
+            weekNumber: weekIndex - 1,
+            dayNumbers: daysArr,
+          });
+        }
+        continue;
+      }
+
+      calendar.push({
+        weekNumber: weekIndex,
+        dayNumbers: daysArr,
+      });
     }
 
     this.calendar = calendar;
   }
 
-  private changeMonth(number) {
-    if (this.monthValue + number < 1) {
-      this.yearValue--;
-      this.monthValue = 12;
-    } else if (this.monthValue + number > 12) {
-      this.yearValue++;
-      this.monthValue = 1;
-    } else {
-      this.monthValue += number;
-    }
-    this.calculateCalendar();
+  private selectTempYear(event: MouseEvent, year: number) {
+    event?.stopPropagation();
+    this.tempYear = year;
   }
 
-  private selectMonth(month: MonthNumbers) {
-    this.monthValue = month;
-    this.yearValue = this.tempYear;
-    this.tempMonth = month;
-
-    this.hostElement.shadowRoot.querySelector('ix-dropdown').show = false;
+  private focusMonth() {
+    this.firstMonthRef.focus();
   }
 
   private infiniteScrollYears() {
@@ -321,168 +463,220 @@ export class DatePicker {
     const maxScroll = this.yearContainerRef.scrollHeight;
     const atTop = scroll === 0;
     const atBottom =
-      scroll + this.yearContainerRef.getBoundingClientRect().height ===
-      maxScroll;
+      Math.round(scroll + this.yearContainerRef.offsetHeight) >= maxScroll;
     const limit = 200;
 
-    if (this.years.length > limit) return;
+    if (this.endYear - this.startYear > limit) return;
 
     if (atTop) {
-      this.years = [
-        ...[...Array(5).keys()].map((year) => year + this.years[0] - 5),
-        ...this.years,
-      ];
-      this.yearContainerRef.scroll({ behavior: 'smooth', top: scroll + 100 });
+      const first = this.yearContainerRef.firstElementChild as HTMLElement;
+      this.startYear -= 5;
+      this.yearContainerRef.scrollTo(0, first.offsetTop);
+
+      return;
     }
 
     if (atBottom) {
-      this.years = [
-        ...this.years,
-        ...[...Array(5).keys()].map(
-          (year) => year + this.years[this.years.length - 1]
-        ),
-      ];
-      this.yearContainerRef.scroll({ behavior: 'smooth', top: scroll - 50 });
+      const last = this.yearContainerRef.lastElementChild as HTMLElement;
+      this.endYear += 5;
+      this.yearContainerRef.scrollTo(0, last.offsetTop);
     }
   }
 
-  private selectTempYear(event: MouseEvent, year: number) {
-    event.stopPropagation();
-    this.tempYear = year;
+  private selectMonth(month: number) {
+    this.selectedMonth = month;
+    this.selectedYear = this.tempYear;
+    this.tempMonth = month;
+
+    this.hostElement.shadowRoot.querySelector('ix-dropdown').show = false;
   }
 
-  private getDayClasses(day: number) {
+  private changeToAdjacentMonth(number: -1 | 1) {
+    if (this.selectedMonth + number < 0) {
+      this.selectedYear--;
+      this.selectedMonth = 11;
+    } else if (this.selectedMonth + number > 11) {
+      this.selectedYear++;
+      this.selectedMonth = 0;
+    } else {
+      this.selectedMonth += number;
+    }
+  }
+
+  private selectDay(selectedDay: number) {
+    const date = DateTime.fromJSDate(
+      new Date(this.selectedYear, this.selectedMonth, selectedDay)
+    );
+
+    if (!this.range || this.currFromDate === undefined) {
+      this.currFromDate = date;
+      this.onDateChange();
+
+      return;
+    }
+
+    // Reset the range selection
+    if (this.currToDate !== undefined) {
+      this.currFromDate = date;
+      this.currToDate = undefined;
+      this.onDateChange();
+
+      return;
+    }
+
+    // Swap from/to if the second date is before the current date
+    if (date < this.currFromDate) {
+      this.currToDate = this.currFromDate;
+      this.currFromDate = date;
+      this.onDateChange();
+
+      return;
+    }
+
+    // Set the range normally
+    this.currToDate = date;
+    this.onDateChange();
+  }
+
+  private onDateChange() {
+    this.getCurrentDate().then((date) => {
+      this.dateChange.emit(date);
+      if (this.range) {
+        this.dateRangeChange.emit(date);
+      }
+    });
+  }
+
+  private getDayClasses(day: number): any {
     if (!day) {
       return;
     }
 
-    const todayLocal = DateTime.local();
-    const dayLocal = DateTime.local(this.yearValue, this.monthValue, day);
-    const dayIso = dayLocal.toISO();
-    const startIso = this.start?.toISO();
-    const endIso = this.end?.toISO();
-    const isToday = Math.ceil(dayLocal.diff(todayLocal, 'days').days) === 0;
+    const todayObj = DateTime.now();
+    const selectedDayObj = DateTime.fromJSDate(
+      new Date(this.selectedYear, this.selectedMonth, day)
+    );
 
     return {
       'calendar-item': true,
       'empty-day': day === undefined,
-      today: isToday,
+      today: todayObj.hasSame(selectedDayObj, 'day'),
       selected:
-        (this.start && dayIso === startIso) || (this.end && dayIso === endIso),
-      range: this.start && this.end && dayIso > startIso && dayIso < endIso,
-      disabled: !this.isWithinMinMax(dayLocal),
+        this.currFromDate?.hasSame(selectedDayObj, 'day') ||
+        this.currToDate?.hasSame(selectedDayObj, 'day'),
+      range:
+        selectedDayObj.startOf('day') > this.currFromDate?.startOf('day') &&
+        this.currToDate !== undefined &&
+        selectedDayObj.startOf('day') < this.currToDate?.startOf('day'),
+      disabled: !this.isWithinMinMaxDate(selectedDayObj),
     };
   }
 
-  private selectDay(day: number) {
-    const date = DateTime.local(this.yearValue, this.monthValue, day);
-    const isStartBeforeEnd = this.start && this.start.toISO() < date.toISO();
-    const isSameDay =
-      this.start && !this.end && this.start.toISO() === date.toISO();
+  private isWithinMinMaxYear(year: number): boolean {
+    const minDateYear = this.minDate
+      ? DateTime.fromFormat(this.minDate, this.format).year
+      : undefined;
+    const maxDateYear = this.maxDate
+      ? DateTime.fromFormat(this.maxDate, this.format).year
+      : undefined;
+    const isBefore = minDateYear ? year < minDateYear : false;
+    const isAfter = maxDateYear ? year > maxDateYear : false;
 
-    if (day === undefined) return;
-
-    if (isSameDay) {
-      this.start = null;
-      this.onDateChange();
-      return;
-    }
-
-    if (this.range) {
-      if (this.start === null) {
-        this.start = date;
-      } else if (this.end === null) {
-        if (isStartBeforeEnd) {
-          this.end = date;
-        } else {
-          this.end = this.start;
-          this.start = date;
-        }
-      } else {
-        this.start = date;
-        this.end = null;
-      }
-    } else {
-      this.start = date;
-    }
-
-    this.onDateChange();
+    return !isBefore && !isAfter;
   }
 
-  private getOutputFormat() {
-    if (!this.start) {
-      return null;
-    }
+  private isWithinMinMaxMonth(month: number): boolean {
+    const minDateObj = this.minDate
+      ? DateTime.fromFormat(this.minDate, this.format)
+      : undefined;
+    const maxDateObj = this.maxDate
+      ? DateTime.fromFormat(this.maxDate, this.format)
+      : undefined;
+    const minDateMonth = minDateObj?.month;
+    const maxDateMonth = maxDateObj?.month;
+    const isBefore = minDateMonth
+      ? this.tempYear === minDateObj.year && month < minDateMonth
+      : false;
+    const isAfter = maxDateMonth
+      ? this.tempYear === maxDateObj.year && month > maxDateMonth
+      : false;
 
-    if (!this.end) {
-      return this.start.toFormat(this.format);
-    }
-
-    return [
-      this.start.toFormat(this.format),
-      this.end.toFormat(this.format),
-    ].join(this.eventDelimiter);
+    return !isBefore && !isAfter;
   }
 
-  private isWithinMinMax(date: DateTime) {
-    const dateIso = date.toISO();
+  private isWithinMinMaxDate(date: DateTime): boolean {
     const _minDate = this.minDate
       ? DateTime.fromFormat(this.minDate, this.format)
-      : null;
+      : undefined;
     const _maxDate = this.maxDate
       ? DateTime.fromFormat(this.maxDate, this.format)
-      : null;
-    return (
-      (!_minDate || _minDate.toISO() <= dateIso) &&
-      (!_maxDate || _maxDate.toISO() >= dateIso)
-    );
+      : undefined;
+    const isBefore = _minDate
+      ? date.startOf('day') < _minDate.startOf('day')
+      : false;
+    const isAfter = _maxDate
+      ? date.startOf('day') > _maxDate.startOf('day')
+      : false;
+
+    return !isBefore && !isAfter;
   }
 
-  componentWillLoad() {
-    if (this.from === null) {
-      this.start = null;
+  private renderYears(): any[] {
+    const rows = [];
+
+    for (let year = this.startYear; year <= this.endYear; year++) {
+      rows.push(
+        <div
+          key={year}
+          class={{
+            arrowYear: true,
+            'month-dropdown-item': true,
+            'disabled-item': !this.isWithinMinMaxYear(year),
+          }}
+          onClick={(event) => this.selectTempYear(event, year)}
+          onKeyUp={(event) => {
+            if (event.key === 'Enter') {
+              this.selectTempYear(null, year);
+              this.focusMonth();
+            }
+          }}
+          tabIndex={0}
+        >
+          <ix-icon
+            class={{
+              hidden: this.tempYear !== year,
+              arrowPosition: true,
+            }}
+            name="chevron-right"
+            size="12"
+          ></ix-icon>
+          <div style={{ 'min-width': 'max-content' }}>{`${year}`}</div>
+        </div>
+      );
     }
 
-    if (this.year !== null) {
-      this.yearValue = this.year;
-    }
-    if (this.month) {
-      this.monthValue = this.month;
-    }
-  }
-
-  componentWillRender() {
-    this.calculateCalendar();
-  }
-
-  /**
-   * Get the current DateTime
-   */
-  @Method()
-  async getCurrentDate() {
-    return {
-      start: this.start?.toFormat(this.format),
-      end: this.end?.toFormat(this.format),
-    };
+    return rows;
   }
 
   render() {
     return (
       <Host>
-        <ix-date-time-card individual={this.individual} corners={this.corners}>
+        <ix-date-time-card
+          corners={this.corners}
+          standaloneAppearance={this.standaloneAppearance}
+        >
           <div class="header" slot="header">
             <ix-icon-button
-              onClick={() => this.changeMonth(-1)}
+              onClick={() => this.changeToAdjacentMonth(-1)}
               ghost
-              icon={'chevron-left'}
+              icon="chevron-left"
               variant="primary"
               class="arrows"
             ></ix-icon-button>
-
             <div class="selector">
               <ix-button ghost ref={(ref) => (this.dropdownButtonRef = ref)}>
                 <span class="fontSize capitalize">
-                  {this.monthNames[this.monthValue - 1]} {this.yearValue}
+                  {this.monthNames[this.selectedMonth]} {this.selectedYear}
                 </span>
               </ix-button>
               <ix-dropdown
@@ -496,50 +690,43 @@ export class DatePicker {
                     onScroll={() => this.infiniteScrollYears()}
                     ref={(ref) => (this.yearContainerRef = ref)}
                   >
-                    {this.years.map((year) => (
-                      <div
-                        key={year}
-                        class={{ arrowYear: true }}
-                        onClick={(event) => this.selectTempYear(event, year)}
-                      >
-                        <ix-icon
-                          class={{
-                            hidden: this.tempYear !== year,
-                            arrowPosition: true,
-                          }}
-                          name={'chevron-right'}
-                          size="12"
-                        ></ix-icon>
-                        <div
-                          style={{ 'min-width': 'max-content' }}
-                        >{`${year}`}</div>
-                      </div>
-                    ))}
+                    {this.renderYears()}
                   </div>
                   <div class="overflow">
                     {this.monthNames.map((month, index) => (
                       <div
                         key={month}
+                        ref={(ref) => {
+                          if (month === this.monthNames[0]) {
+                            this.firstMonthRef = ref as HTMLElement;
+                          }
+                        }}
                         class={{
                           arrowYear: true,
-                          selected: this.tempMonth - 1 === index,
+                          'month-dropdown-item': true,
+                          selected:
+                            this.tempYear === this.selectedYear &&
+                            this.tempMonth === index,
+                          'disabled-item': !this.isWithinMinMaxMonth(index),
                         }}
-                        onClick={() =>
-                          this.selectMonth((index + 1) as MonthNumbers)
+                        onClick={() => this.selectMonth(index)}
+                        onKeyUp={(event) =>
+                          event.key === 'Enter' && this.selectMonth(index)
                         }
+                        tabIndex={0}
                       >
                         <ix-icon
                           class={{
-                            hidden: this.tempMonth - 1 !== index,
+                            hidden:
+                              this.tempYear !== this.selectedYear ||
+                              this.tempMonth !== index,
                             checkPosition: true,
                           }}
-                          name={'single-check'}
+                          name="single-check"
                           size="16"
                         ></ix-icon>
                         <div>
-                          <span
-                            class={{ capitalize: true, monthMargin: true }}
-                          >{`${month} ${this.tempYear}`}</span>
+                          <span class="capitalize monthMargin">{`${month} ${this.tempYear}`}</span>
                         </div>
                       </div>
                     ))}
@@ -547,16 +734,14 @@ export class DatePicker {
                 </div>
               </ix-dropdown>
             </div>
-
             <ix-icon-button
-              onClick={() => this.changeMonth(1)}
+              onClick={() => this.changeToAdjacentMonth(1)}
               ghost
-              icon={'chevron-right'}
+              icon="chevron-right"
               variant="primary"
               class="arrows"
             ></ix-icon-button>
           </div>
-
           <div class="grid">
             <div class="calendar-item week-day"></div>
             {this.dayNames.map((name) => (
@@ -564,16 +749,21 @@ export class DatePicker {
                 {name.slice(0, 3)}
               </div>
             ))}
-
             {this.calendar.map((week) => {
               return (
                 <Fragment>
-                  <div class="calendar-item week-number">{week[0]}</div>
-                  {week[1].map((day) => (
+                  <div class="calendar-item week-number">{week.weekNumber}</div>
+                  {week.dayNumbers.map((day) => (
                     <div
                       key={day}
+                      id={`day-cell-${day}`}
+                      date-calender-day
                       class={this.getDayClasses(day)}
                       onClick={() => this.selectDay(day)}
+                      onKeyUp={(e) => e.key === 'Enter' && this.selectDay(day)}
+                      tabIndex={day === this.focusedDay ? 0 : -1}
+                      onFocus={() => this.onDayFocus()}
+                      onBlur={() => this.onDayBlur()}
                     >
                       {day}
                     </div>
@@ -582,9 +772,11 @@ export class DatePicker {
               );
             })}
           </div>
-
           <div
-            class={{ button: true, hidden: !this.individual || !this.range }}
+            class={{
+              button: true,
+              hidden: !this.range || !this.standaloneAppearance,
+            }}
           >
             <ix-button onClick={() => this.onDone()}>
               {this.textSelectDate}
