--- packages/core/src/components/time-picker-rework/time-picker-rework.scss
@@ -1,118 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2023 Siemens AG
- *
- * SPDX-License-Identifier: MIT
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-@import 'common-variables';
-@import 'mixins/shadow-dom/component';
-@import 'components/form/input';
-
-:host {
-  @include ix-component;
-  @import 'legacy/components/forms';
-
-  display: block;
-  position: relative;
-  width: 22rem;
-
-  @media (max-width: 576px) {
-    .clock {
-      padding: 1.5rem 0;
-    }
-
-    .header {
-      height: $medium-control-height !important;
-    }
-  }
-
-  @media (min-width: 576px) {
-    .clock {
-      padding: 4.25rem 0;
-    }
-
-    .header {
-      height: $large-control-height !important;
-    }
-  }
-
-  .header {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: $large-control-height;
-  }
-
-  .flex {
-    display: flex;
-  }
-
-  .clock {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-
-  input,
-  .time-reference {
-    width: 2.625rem;
-    height: $large-space;
-    text-align: center;
-    margin-top: 0.25rem;
-    margin-bottom: 0.25rem;
-
-    /* Chrome, Safari, Edge, Opera */
-    &::-webkit-outer-spin-button,
-    &::-webkit-inner-spin-button {
-      -webkit-appearance: none;
-      margin: 0 0.3125rem;
-    }
-
-    /* Firefox */
-    &[type='number'] {
-      -moz-appearance: textfield;
-    }
-  }
-
-  .form-control[type='number'] {
-    text-align: center;
-  }
-
-  .columns {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-around;
-  }
-
-  .column-seperator {
-    display: flex;
-    align-items: center;
-    margin: 0 $tiny-space;
-  }
-
-  .button {
-    display: flex;
-    justify-content: flex-end;
-  }
-
-  .default-space {
-    margin-left: $default-space;
-  }
-
-  .text-align {
-    text-align: center;
-  }
-
-  .hidden {
-    display: none;
-  }
-
-  .time-reference {
-    margin-top: 0.5rem;
-    margin-bottom: 0.5rem;
-  }
-}
