+++ packages/core/src/components/time-picker/time-picker.tsx
@@ -16,12 +16,28 @@ import {
   Method,
   Prop,
   State,
+  Watch,
 } from '@stencil/core';
 import { DateTime } from 'luxon';
 import { DateTimeCardCorners } from '../date-time-card/date-time-card';
 
 export type TimePickerCorners = DateTimeCardCorners;
 
+interface TimePickerDescriptor {
+  unit: 'hour' | 'minute' | 'second';
+  placeholder: string;
+  hidden: boolean;
+}
+
+interface TimeOutputFormat {
+  hour: string;
+  minute: string;
+  second: string;
+}
+
+/**
+ * @internal
+ */
 @Component({
   tag: 'ix-time-picker',
   styleUrl: 'time-picker.scss',
@@ -30,6 +46,7 @@ export type TimePickerCorners = DateTimeCardCorners;
 export class TimePicker {
   /**
    * Format of time string
+   * See {@link "https://moment.github.io/luxon/#/formatting?id=table-of-tokens"} for all available tokens.
    *
    * @since 1.1.0
    */
@@ -41,46 +58,59 @@ export class TimePicker {
   @Prop() corners: TimePickerCorners = 'rounded';
 
   /**
-   * @deprecated Will be removed in 2.0.0
+   * Controls the visual presentation and styling of the component when it is displayed as a standalone element
+   */
+  @Prop() standaloneAppearance: boolean = true;
+
+  /**
+   * @deprecated Not supported since 2.0.0.
    */
   @Prop() individual: boolean = true;
 
   /**
    * Show hour input
    */
-  @Prop() showHour = false;
+  @Prop() showHour = true;
 
   /**
    * Show minutes input
    */
-  @Prop() showMinutes = false;
+  @Prop() showMinutes = true;
 
   /**
    * Show seconds input
    */
-  @Prop() showSeconds = false;
+  @Prop() showSeconds = true;
 
   /**
    * Select time with format string
+   * Format has to match the `format` property.
    *
    * @since 1.1.0
    */
   @Prop() time: string = DateTime.now().toFormat(this.format);
 
+  @Watch('time')
+  watchTimePropHandler(newValue: string) {
+    this._time = DateTime.fromFormat(newValue, this.format);
+    if (!this._time.isValid) {
+      throw new Error('Format is not supported or not correct');
+    }
+  }
+
   /**
    * Show time reference input
    *
-   * @since 1.1.0 time reference is default aligned with formt tt
+   * @since 1.1.0 time reference is default aligned with format tt
+   * @deprecated Since 2.0.0 time reference will be displayed depending on format.
    */
-  @Prop({ mutable: true }) showTimeReference = undefined;
+  // eslint-disable-next-line @stencil-community/strict-mutable
+  @Prop({ mutable: true }) showTimeReference?: boolean = undefined;
 
   /**
    * Set time reference
    */
-  @Prop({ mutable: true }) timeReference: 'AM' | 'PM' = DateTime.fromFormat(
-    this.time,
-    this.format
-  ).toFormat('a') as 'PM' | 'AM';
+  @Prop() timeReference: 'AM' | 'PM' | undefined;
 
   /**
    * Text of date select button
@@ -90,7 +120,18 @@ export class TimePicker {
   @Prop() textSelectTime = 'Done';
 
   /**
+   * Text for top label
+   */
+  @Prop() textTime: string = 'Time';
+
+  /**
    * Time event
+   */
+  @Event() timeSelect: EventEmitter<string>;
+
+  /**
+   * Time event
+   * @deprecated Will be removed in 3.0.0. Use `time-select` event.
    */
   @Event() done: EventEmitter<string>;
 
@@ -99,366 +140,222 @@ export class TimePicker {
    */
   @Event() timeChange: EventEmitter<string>;
 
-  @State() hourInputRef: HTMLInputElement;
-  @State() minuteInputRef: HTMLInputElement;
-  @State() secondInputRef: HTMLInputElement;
-  @State() referenceInputRef: HTMLInputElement;
-
-  get hour() {
-    return this._time.hour;
-  }
-
-  get minutes() {
-    return this._time.minute;
-  }
-
-  get seconds() {
-    return this._time.second;
-  }
-
-  private _time: DateTime = DateTime.now();
-
-  private setInputValue(
-    inputElement: HTMLInputElement,
-    value: number,
-    max: number
-  ) {
-    if (value > max) {
-      inputElement.value = max.toString();
-    } else if (value < 0) {
-      inputElement.value = '0';
-    } else {
-      inputElement.value = value.toString();
-    }
-
-    this.updateAndEmitTime();
-  }
-
-  private updateAndEmitTime() {
-    this._time = this._time.set({
-      hour: Number(this.hourInputRef.value),
-      minute: Number(this.minuteInputRef.value),
-      second: Number(this.secondInputRef.value),
-    });
-
-    this.setHourAccordingToReference();
-
-    this.emitTimeChange();
-  }
-
-  private toggleInputValue(
-    inputElement: HTMLInputElement,
-    currentValue: number,
-    step: 'up' | 'down',
-    max: number
-  ) {
-    if (step === 'up') {
-      if (currentValue === max) {
-        inputElement.value = '0';
-      } else {
-        inputElement.stepUp();
-      }
-    } else {
-      if (currentValue === 0) {
-        inputElement.value = max.toString();
-      } else {
-        inputElement.stepDown();
-      }
-    }
-  }
-
-  private toggleHourInputWithRef(
-    inputElement: HTMLInputElement,
-    currentValue: number,
-    step: 'up' | 'down'
-  ) {
-    if (step === 'up') {
-      if (currentValue === this.getMaxHour()) {
-        inputElement.value = '0';
-      } else {
-        inputElement.stepUp();
-      }
-    } else {
-      if (
-        (this.showTimeReference &&
-          this.timeReference === 'PM' &&
-          currentValue === 12) ||
-        currentValue === 0
-      ) {
-        inputElement.value = this.getMaxDisplayedHour().toString();
-      } else {
-        inputElement.stepDown();
-      }
-    }
-  }
-
-  private updateInput(
-    step: 'up' | 'down',
-    { hours = undefined, minutes = undefined, seconds = undefined }
-  ) {
-    if (hours) {
-      if (this.showTimeReference) {
-        this.toggleHourInputWithRef(this.hourInputRef, this.hour, step);
-      } else {
-        this.toggleInputValue(this.hourInputRef, this.hour, step, 23);
-      }
-    }
-
-    if (minutes) {
-      this.toggleInputValue(this.minuteInputRef, this.minutes, step, 59);
-    }
-
-    if (seconds) {
-      this.toggleInputValue(this.secondInputRef, this.seconds, step, 59);
-    }
-
-    this.updateAndEmitTime();
-  }
-
-  private changeReference() {
-    if (this.timeReference === 'AM') {
-      this.timeReference = 'PM';
-    } else {
-      this.timeReference = 'AM';
-    }
-
-    this.setHourAccordingToReference();
-
-    this.emitTimeChange();
-  }
-
-  private setHourAccordingToReference() {
-    if (!this.showTimeReference) {
-      return;
-    }
-
-    let hour = Number(this.hourInputRef.value);
-
-    if (this.timeReference === 'PM') hour += 12;
-
-    this._time = this._time.set({ hour });
-  }
-
-  private emitTimeChange() {
-    const time = this._time.toFormat(this.format);
-    this.timeChange.emit(time);
-  }
+  @State() private _time: DateTime;
+  @State() private _timeRef: 'AM' | 'PM' | undefined;
+  @State() private _formattedTime: TimeOutputFormat;
 
   componentWillLoad() {
     this._time = DateTime.fromFormat(this.time, this.format);
-    if (this.showTimeReference === undefined) {
-      const matchedKeys = Object.keys(
-        DateTime.fromFormatExplain(this.time, this.format).matches
+
+    if (!this._time.isValid) {
+      console.error(
+        `Invalid time format. The configured format does not match the format of the passed time. ${this._time.invalidReason}: ${this._time.invalidExplanation}`
       );
-      this.showTimeReference = matchedKeys.includes('a');
+      return;
+    }
+
+    this._timeRef = this.format.includes('a')
+      ? (DateTime.fromFormat(this.time, this.format).toFormat('a') as
+          | 'AM'
+          | 'PM')
+      : undefined;
+    this.formatTime();
+  }
+
+  @Watch('_time')
+  formatTime() {
+    const [hour, minute, second] = this._time
+      .toFormat(this.format)
+      .split(' ')[0]
+      .split(':');
+
+    this._formattedTime = {
+      hour: hour,
+      minute: minute,
+      second: second,
+    };
+  }
+
+  @Watch('_time')
+  onInternalTimeChange() {
+    this.timeChange.emit(this._time.toFormat(this.format));
+    if (this._timeRef) this._timeRef = this._time.toFormat('a') as 'AM' | 'PM';
+  }
+
+  timeUpdate(unit: 'hour' | 'minute' | 'second', value: number): number {
+    let maxValue = DateTime.now().endOf('day').get(unit);
+
+    if (this._timeRef === 'PM' && unit === 'hour') value += 12;
+    if (this._timeRef === 'AM' && unit === 'hour') maxValue = 12;
+
+    if (value > maxValue) {
+      value = maxValue;
+    } else if (value < 0) {
+      value = 0;
+    }
+
+    this._time = this._time.set({
+      [unit]: value,
+    });
+    return value;
+  }
+
+  changeTimeReference() {
+    this._timeRef = this._timeRef === 'AM' ? 'PM' : 'AM';
+
+    if (!this._time.toFormat('a').includes(this._timeRef)) {
+      this._time = this._time.plus({
+        hour: 12,
+      });
     }
   }
 
   /**
-   * Get current time
+   * Get the current time based on the wanted format
    */
   @Method()
   async getCurrentTime() {
     return this._time.toFormat(this.format);
   }
 
-  private getDisplayedHour() {
-    if (this.showTimeReference && this.timeReference === 'PM') {
-      return this._time.hour - 12;
-    }
-
-    return this._time.hour;
-  }
-
-  private getMaxDisplayedHour() {
-    if (this.showTimeReference) {
-      return 11;
-    }
-
-    return 23;
-  }
-
-  private getMaxHour() {
-    if (this.showTimeReference && this.timeReference === 'AM') {
-      return 11;
-    }
-
-    return 23;
-  }
-
   render() {
-    let hideHour = !this.showHour;
-    let hideMinutes = !this.showMinutes;
-    let hideSeconds = !this.showSeconds;
-    const hideTimeReference = !this.showTimeReference;
-    const hideIndividual = !this.individual;
+    let timepickerInformation: TimePickerDescriptor[] = [
+      {
+        unit: 'hour',
+        placeholder: 'HH',
+        hidden: !this.showHour,
+      },
+      {
+        unit: 'minute',
+        placeholder: 'MM',
+        hidden: !this.showMinutes,
+      },
+      {
+        unit: 'second',
+        placeholder: 'SS',
+        hidden: !this.showSeconds,
+      },
+    ];
 
-    if (hideHour && hideMinutes && hideSeconds) {
-      hideHour = false;
-      hideMinutes = false;
-      hideSeconds = false;
-    }
-
-    const hideHourSeperator = hideMinutes || hideHour;
-    const hideMinutesSeperator = hideSeconds || hideMinutes;
+    timepickerInformation = timepickerInformation.filter(
+      (item) => !item.hidden
+    );
 
     return (
       <Host>
-        <ix-date-time-card individual={this.individual} corners={this.corners}>
+        <ix-date-time-card
+          standaloneAppearance={this.standaloneAppearance}
+          corners={this.corners}
+        >
           <div class="header" slot="header">
-            <ix-typography variant="default-title">Time</ix-typography>
+            <ix-typography variant="default-title">
+              {this.textTime || 'Time'}
+            </ix-typography>
           </div>
-
           <div class="clock">
-            <div class={{ columns: true, hidden: hideHour }}>
-              <ix-icon-button
-                size="16"
-                onClick={() => this.updateInput('up', { hours: true })}
-                ghost
-                icon={'chevron-up'}
-                variant="primary"
-                class="arrows"
-              ></ix-icon-button>
-              <input
-                class="form-control"
-                name="hours"
-                type="number"
-                placeholder="HH"
-                min="0"
-                max={this.getMaxDisplayedHour()}
-                value={this.getDisplayedHour()}
-                ref={(ref) => (this.hourInputRef = ref)}
-                onChange={() =>
-                  this.setInputValue(
-                    this.hourInputRef,
-                    Number(this.hourInputRef.value),
-                    this.getMaxDisplayedHour()
-                  )
-                }
-              ></input>
-              <ix-icon-button
-                size="16"
-                onClick={() => this.updateInput('down', { hours: true })}
-                ghost
-                icon={'chevron-down'}
-                variant="primary"
-                class="arrows"
-              ></ix-icon-button>
-            </div>
+            {timepickerInformation.map((descriptor, index: number) => (
+              <div class="flex">
+                <div class={{ columns: true, hidden: descriptor.hidden }}>
+                  <ix-icon-button
+                    size="16"
+                    onClick={() =>
+                      (this._time = this._time.plus({
+                        [descriptor.unit]: 1,
+                      }))
+                    }
+                    ghost
+                    icon="chevron-up"
+                    variant="primary"
+                    class="arrows"
+                  ></ix-icon-button>
 
-            <div
-              class={{ 'column-seperator': true, hidden: hideHourSeperator }}
-            >
-              :
-            </div>
+                  <input
+                    class="form-control"
+                    name={descriptor.unit}
+                    type="number"
+                    placeholder={descriptor.placeholder}
+                    value={
+                      this._formattedTime
+                        ? this._formattedTime[descriptor.unit]
+                        : null
+                    }
+                    onKeyDown={(e) => {
+                      if (e.key !== 'ArrowUp' && e.key !== 'ArrowDown') return;
 
-            <div class={{ columns: true, hidden: hideMinutes }}>
-              <ix-icon-button
-                size="16"
-                onClick={() => this.updateInput('up', { minutes: true })}
-                ghost
-                icon={'chevron-up'}
-                variant="primary"
-                class="arrows"
-              ></ix-icon-button>
-              <input
-                class="form-control"
-                name="minutes"
-                type="number"
-                placeholder="MM"
-                min="0"
-                max="59"
-                value={this.minutes}
-                ref={(ref) => (this.minuteInputRef = ref)}
-                onChange={() =>
-                  this.setInputValue(
-                    this.minuteInputRef,
-                    Number(this.minuteInputRef.value),
-                    59
-                  )
-                }
-              ></input>
-              <ix-icon-button
-                size="16"
-                onClick={() => this.updateInput('down', { minutes: true })}
-                ghost
-                icon={'chevron-down'}
-                variant="primary"
-                class="arrows"
-              ></ix-icon-button>
-            </div>
+                      const value = e.key === 'ArrowUp' ? 1 : -1;
+                      this._time = this._time.plus({
+                        [descriptor.unit]: value,
+                      });
+                      e.preventDefault();
+                    }}
+                    onChange={(e: any) => {
+                      let inputElement = e.target as HTMLInputElement;
+                      inputElement.value = this.timeUpdate(
+                        descriptor.unit,
+                        +inputElement.value
+                      ).toString();
+                    }}
+                  ></input>
 
-            <div
-              class={{ 'column-seperator': true, hidden: hideMinutesSeperator }}
-            >
-              :
-            </div>
+                  <ix-icon-button
+                    size="16"
+                    onClick={() =>
+                      (this._time = this._time.minus({
+                        [descriptor.unit]: 1,
+                      }))
+                    }
+                    ghost
+                    icon="chevron-down"
+                    variant="primary"
+                    class="arrows"
+                  ></ix-icon-button>
+                </div>
 
-            <div class={{ columns: true, hidden: hideSeconds }}>
-              <ix-icon-button
-                size="16"
-                onClick={() => this.updateInput('up', { seconds: true })}
-                ghost
-                icon={'chevron-up'}
-                variant="primary"
-                class="arrows"
-              ></ix-icon-button>
-              <input
-                class="form-control"
-                name="seconds"
-                type="number"
-                placeholder="SS"
-                min="0"
-                max="59"
-                value={this.seconds}
-                ref={(ref) => (this.secondInputRef = ref)}
-                onChange={() =>
-                  this.setInputValue(
-                    this.secondInputRef,
-                    Number(this.secondInputRef.value),
-                    59
-                  )
-                }
-              ></input>
-              <ix-icon-button
-                size="16"
-                onClick={() => this.updateInput('down', { seconds: true })}
-                ghost
-                icon={'chevron-down'}
-                variant="primary"
-                class="arrows"
-              ></ix-icon-button>
-            </div>
+                {index !== timepickerInformation.length - 1 && (
+                  <div
+                    class={{
+                      'column-seperator': true,
+                      hidden: descriptor.hidden,
+                    }}
+                  >
+                    :
+                  </div>
+                )}
+              </div>
+            ))}
 
             <div
               class={{
                 columns: true,
                 'default-space': true,
-                hidden: hideTimeReference,
+                hidden: this._timeRef === undefined,
               }}
             >
               <ix-icon-button
                 size="16"
-                onClick={() => this.changeReference()}
+                onClick={() => this.changeTimeReference()}
                 ghost
-                icon={'chevron-up'}
+                icon="chevron-up"
                 variant="primary"
                 class="arrows"
               ></ix-icon-button>
-              <div class="time-reference">{this.timeReference}</div>
+              <div class="time-reference">{this._timeRef}</div>
               <ix-icon-button
                 size="16"
-                onClick={() => this.changeReference()}
+                onClick={() => this.changeTimeReference()}
                 ghost
-                icon={'chevron-down'}
+                icon="chevron-down"
                 variant="primary"
                 class="arrows"
               ></ix-icon-button>
             </div>
           </div>
-          <div class={{ button: true, hidden: hideIndividual }}>
+          <div class={{ button: true, hidden: !this.standaloneAppearance }}>
             <ix-button
-              onClick={() => this.done.emit(this._time.toFormat(this.format))}
+              onClick={() => {
+                this.timeSelect.emit(this._time.toFormat(this.format));
+                this.done.emit(this._time.toFormat(this.format));
+              }}
             >
               {this.textSelectTime}
             </ix-button>
