+++ packages/core/src/tests/date-picker-rework/basic/index.html
@@ -22,12 +22,12 @@ SPDX-License-Identifier: MIT
         margin-top: 2rem;
       "
     >
-      <ix-date-picker-rework
+      <ix-date-picker
         range="false"
         from="1990/03/29"
         min-date="1990/03/05"
         max-date="1990/03/30"
-      ></ix-date-picker-rework>
+      ></ix-date-picker>
     </div>
     <script src="http://127.0.0.1:8080/scripts/e2e/load-e2e-runtime.js"></script>
   </body>
