+++ packages/core/src/tests/date-picker-rework/week-start/index.html
@@ -22,11 +22,11 @@ SPDX-License-Identifier: MIT
         margin-top: 2rem;
       "
     >
-      <ix-date-picker-rework
+      <ix-date-picker
         range="false"
         from="2023/08/15"
         week-start-index="1"
-      ></ix-date-picker-rework>
+      ></ix-date-picker>
     </div>
     <script src="http://127.0.0.1:8080/scripts/e2e/load-e2e-runtime.js"></script>
   </body>
