+++ packages/core/src/tests/date-picker-rework/year-month-selection/index.html
@@ -22,12 +22,12 @@ SPDX-License-Identifier: MIT
         margin-top: 2rem;
       "
     >
-      <ix-date-picker-rework
+      <ix-date-picker
         range="false"
         from="2023/08/15"
         min-date="2023/05/01"
         max-date="2023/09/25"
-      ></ix-date-picker-rework>
+      ></ix-date-picker>
     </div>
     <script src="http://127.0.0.1:8080/scripts/e2e/load-e2e-runtime.js"></script>
   </body>
