+++ packages/core/src/tests/time-picker-rework/basic/index.html
@@ -13,7 +13,7 @@ SPDX-License-Identifier: MIT
     />
     <title>Stencil Component Starter</title>
     <style>
-      ix-time-picker-rework {
+      ix-time-picker {
         margin-inline-end: 1rem;
       }
     </style>
@@ -27,15 +27,15 @@ SPDX-License-Identifier: MIT
         margin-top: 2rem;
       "
     >
-      <ix-time-picker-rework style="width: min-content" time="10:11:12"></ix-time-picker-rework>
-      <ix-time-picker-rework style="width: min-content" time="10:11:12" show-minutes="false" show-seconds="false"></ix-time-picker-rework>
-      <ix-time-picker-rework style="width: min-content" time="10:11:12" show-hours="false" show-seconds="false"></ix-time-picker-rework>
-      <ix-time-picker-rework style="width: min-content" time="10:11:12" show-hours="false" show-minutes="false"></ix-time-picker-rework>
-      <ix-time-picker-rework
+      <ix-time-picker style="width: min-content" time="10:11:12"></ix-time-picker>
+      <ix-time-picker style="width: min-content" time="10:11:12" show-minutes="false" show-seconds="false"></ix-time-picker>
+      <ix-time-picker style="width: min-content" time="10:11:12" show-hours="false" show-seconds="false"></ix-time-picker>
+      <ix-time-picker style="width: min-content" time="10:11:12" show-hours="false" show-minutes="false"></ix-time-picker>
+      <ix-time-picker
         style="width: min-content"
         time="10:11:12 AM"
         format="hh:mm:ss a"
-      ></ix-time-picker-rework>
+      ></ix-time-picker>
     </div>
     <script src="http://127.0.0.1:8080/scripts/e2e/load-e2e-runtime.js"></script>
   </body>
